requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
requests==2.31.0
python-dotenv==1.0.0 